+++ events/messageReactionAdd.js
@@ -1,4 +1,4 @@
-import { EmbedBuilder } from "discord.js";
+import getStarboardMessage from "../util/getStarboardMessage.js";
 import logger from "../util/logger.js";
 import globalVars from "../objects/globalVars.json" with { type: "json" };
 
@@ -7,6 +7,7 @@ const altboardChannelID = "1234922298255872092"; // Evil starboard
 const altboardEmote = "<:nostar:780198211913646130>";
 const altboardEmoteID = altboardEmote.replace(/[^0-9]+/g, "");
 
+// VIDEO ???
 export default async (client, messageReaction) => {
     try {
         let boardEmote = starboardEmote;
@@ -54,58 +55,11 @@ export default async (client, messageReaction) => {
         } else {
             starLimit = globalVars.starboardLimit;
         };
-        // Check for atached files
-        let messageImage = null;
-        let starboardEmbeds = [];
-        // let seperateFiles = null;
-        if (targetMessage.attachments.size > 0) {
-            messageImage = targetMessage.attachments.first().proxyURL;
-            // Videos can't be embedded unless you're X (formerly Twitter) or YouTube, so they are sent as seperate mesages
-            // if (messageImage.endsWith(".mp4")) seperateFiles = messageImage;
-            targetMessage.attachments.forEach(attachment => {
-                if (attachment.proxyURL !== messageImage) {
-                    let imageEmbed = new EmbedBuilder()
-                        .setImage(attachment.proxyURL)
-                        .setURL(targetMessage.url);
-                    starboardEmbeds.push(imageEmbed);
-                };
-            });
-        };
-        // Get user's avatar, try to use server avatar, otherwise default to global avatar
-        let avatar = targetMessage.author.displayAvatarURL(globalVars.displayAvatarSettings);
-        if (targetMessage.member) avatar = targetMessage.member.displayAvatarURL(globalVars.displayAvatarSettings);
-        // Check if the starred message is replying to another message
-        let isReply = false;
-        let replyMessage = null;
-        let replyString = "";
-        if (targetMessage.reference) isReply = true;
-        if (isReply) {
-            try {
-                // Format message the starred message is replying to
-                replyMessage = await targetMessage.channel.messages.fetch(targetMessage.reference.messageId);
-                if (replyMessage.content.length > 0) replyString += `"${replyMessage.content.slice(0, 950)}"`;
-                if (replyMessage.author) replyString += `\n-${replyMessage.author}`;
-            } catch (e) {
-                isReply = false;
-            };
-        };
-        // Format the starboard embed message
-        const starEmbed = new EmbedBuilder()
-            .setColor(globalVars.embedColor)
-            .setTitle(`${boardEmote}${messageReaction.count}`)
-            .setURL(targetMessage.url)
-            .setThumbnail(avatar)
-            .setImage(messageImage)
-            .setFooter({ text: targetMessage.author.username })
-            .setTimestamp(targetMessage.createdTimestamp);
-        if (targetMessage.content) starEmbed.setDescription(targetMessage.content);
-        starEmbed.addFields([{ name: `Sent:`, value: `By ${targetMessage.author} in ${targetMessage.channel}\nContext: ${targetMessage.url}`, inline: false }]);
-        if (isReply && replyString.length > 0) starEmbed.addFields([{ name: `Replying to:`, value: replyString, inline: true }]);
-        starboardEmbeds.unshift(starEmbed);
+        let starboardMessage = await getStarboardMessage({ messageReaction: messageReaction, targetMessage: targetMessage, boardEmote: boardEmote });
         // Check if message already existed in database (was posted to starboard) or if star amount simply changed
         if (messageReaction.count >= starLimit && !messageDB) {
             // Send message then push data to database
-            await starboard.send({ embeds: starboardEmbeds }).then(async (m) => await serverApi.StarboardMessages.upsert({ channel_id: targetMessage.channel.id, message_id: targetMessage.id, starboard_channel_id: m.channel.id, starboard_message_id: m.id }));
+            await starboard.send(starboardMessage).then(async (m) => await serverApi.StarboardMessages.upsert({ channel_id: targetMessage.channel.id, message_id: targetMessage.id, starboard_channel_id: m.channel.id, starboard_message_id: m.id }));
             return;
         } else if (messageDB) {
             // Update existing starboard message and database entry
@@ -113,7 +67,7 @@ export default async (client, messageReaction) => {
             let starMessage = await starChannel.messages.fetch(messageDB.starboard_message_id);
             if (!starMessage) return;
             if (starChannel !== starboard) return; // Fix cross-updating between starboard and evil starboard
-            await starMessage.edit({ embeds: starboardEmbeds });
+            await starMessage.edit(starboardMessage);
             // Try to pin messages with double stars
             if (messageReaction.count >= starLimit * 2) starMessage.pin().catch(e => {
                 // console.log(e);
